Support locally installed gulp
Gulp usually installed globally so it's available as `gulp`. But in some cases, it may installed locally in `node_modules/gulp/bin/gulp.js`. This change is to support both globally and locally installed gulp.

System will try to execute global `gulp`. When it's failed, it will try to execute local `gulp`.